Refactored Playwright tests

+++ PlaywrightProject/src/test/java/testngtests/AdminPanelTest.java
@@ -1,19 +1,23 @@
 package testngtests;
 
 import com.mailosaur.MailosaurException;
+import listeners.TestListener;
 import models.User;
-import models.testngpages.AdminPanelPage;
+import models.testngpages.adminpanel.AdminPanelPage;
 import org.assertj.core.api.Assertions;
+import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
 import testngtests.abstractclasses.AbstractTest;
 
 import java.io.IOException;
 
+import static models.enums.SystemAlerts.*;
+
+@Listeners(TestListener.class)
 public class AdminPanelTest extends AbstractTest {
 
     private User newUser = new User.UserBuilder().build();
-    private static final String SUCCESS_ALERT_PROMOTING_USER_TEXT = "Success! User with email '%s' has been promoted to ADMIN role.";
-    private static final String SUCCESS_ALERT_DEMOTING_USER_TEXT = "Success! User with email '%s' has been demoted to USER role.";
+
     private AdminPanelPage adminPanelPage;
 
     @Test
@@ -22,41 +26,71 @@ public class AdminPanelTest extends AbstractTest {
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
-        Assertions.assertThat(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
-        Assertions.assertThat(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail())).isFalse();
-        Assertions.assertThat(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
+        assertThatPromoteButtonIsAvailableOrNot(true);
+        assertThatDemoteButtonIsAvailableOrNot(false);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         adminPanelPage.promoteUserByEmail(newUser.getEmail());
-        Assertions.assertThat(adminPanelPage.getTextFromSuccessAlert()).contains(String.format(SUCCESS_ALERT_PROMOTING_USER_TEXT, newUser.getEmail()));
-        Assertions.assertThat(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail())).isFalse();
-        Assertions.assertThat(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
-        Assertions.assertThat(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
+        assertThatSuccessAlertHasExpectedText(String.format(SUCCESS_ALERT_PROMOTING_USER_TEXT.getAlertText(), newUser.getEmail()));
+        assertThatPromoteButtonIsAvailableOrNot(false);
+        assertThatDemoteButtonIsAvailableOrNot(true);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(newUser);
-        Assertions.assertThat(adminPanelPage.isAdminPanelButtonVisible()).isTrue();
+        assertThatAdminPanelButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
         adminPanelPage.demoteUserByEmail(newUser.getEmail());
-        Assertions.assertThat(adminPanelPage.getTextFromSuccessAlert()).contains(String.format(SUCCESS_ALERT_DEMOTING_USER_TEXT, newUser.getEmail()));
-        Assertions.assertThat(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
-        Assertions.assertThat(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail())).isFalse();
-        Assertions.assertThat(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail())).isTrue();
+        assertThatSuccessAlertHasExpectedText(String.format(SUCCESS_ALERT_DEMOTING_USER_TEXT.getAlertText(), newUser.getEmail()));
+        assertThatPromoteButtonIsAvailableOrNot(true);
+        assertThatDemoteButtonIsAvailableOrNot(false);
+        assertThatDeleteButtonIsAvailableOrNot(true);
         signOut();
 
         signIn(newUser);
-        Assertions.assertThat(adminPanelPage.isAdminPanelButtonVisible()).isFalse();
+        assertThatAdminPanelButtonIsAvailableOrNot(false);
         signOut();
 
         signIn(existedAdminUser);
         adminPanelPage = mainPage.goToAdminPanel();
         adminPanelPage.deleteUserByEmail(newUser.getEmail());
-        Assertions.assertThat(adminPanelPage.getTextFromSuccessAlert()).contains(SUCCESS_ALERT_USER_REMOVED_TEXT);
-        Assertions.assertThat(adminPanelPage.isRowForUserWithEmailExist(newUser.getEmail())).isFalse();
+        assertThatSuccessAlertHasExpectedText(SUCCESS_ALERT_USER_REMOVED_TEXT.getAlertText());
+        assertThatRowWithUserIsNotExist(newUser);
         signOut();
 
         signIn(newUser, true);
+    }
+
+    private void assertThatPromoteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isPromoteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Promote button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatDemoteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isDemoteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Demote button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatDeleteButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isDeleteButtonAvailableForUserWithEmail(newUser.getEmail()))
+                .as("Assert that Delete button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatAdminPanelButtonIsAvailableOrNot(boolean shouldBeAvailable) {
+        Assertions.assertThat(adminPanelPage.isAdminPanelButtonVisible())
+                .as("Assert that Admin panel button is available: " + shouldBeAvailable)
+                .isEqualTo(shouldBeAvailable);
+    }
+
+    private void assertThatRowWithUserIsNotExist(User user) {
+        Assertions.assertThat(adminPanelPage.isRowForUserWithEmailExist(user.getEmail()))
+                .as("Assert that row for expected user is not exist")
+                .isFalse();
     }
 
 }
